requirements.txt
flake8==7.1.1
matplotlib==3.9.2
pillow==11.0.0
plantcv==4.5.1
rembg==2.0.59
rembg==2.0.59
numpy==1.26.4
tensorflow-cpu==2.18.0
a==1.0